requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
apex==0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
cmake==3.28.1
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cubinlinker==0.3.0+2.g405ac64
cuda-python==12.3.0rc4+9.gdb8c48a.dirty
cudf==23.12.0
cugraph==23.12.0
cugraph-dgl==23.12.0
cugraph-service-client==23.12.0
cugraph-service-server==23.12.0
cuml==23.12.0
cupy-cuda12x==12.3.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
dask==2023.11.0
dask-cuda==23.12.0
dask-cudf==23.12.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distributed==2023.11.0
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
expecttest==0.1.3
fastjsonschema==2.19.1
fastrlock==0.8.2
filelock==3.13.1
flash-attn==2.0.4
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
geomloss==0.2.6
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.26.2
google-auth-oauthlib==0.4.6
graphsurgeon==0.4.6
grpcio==1.60.0
hydra-core==1.3.2
hypothesis==5.35.1
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.29.0
ipython==8.20.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter-tensorboard==0.2.0
jupyterlab==2.3.2
jupyterlab_pygments==0.3.0
jupyterlab-server==1.2.0
jupytext==1.16.1
keopscore==2.2.3
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
lightning==2.4.0
lightning-utilities==0.11.6
llvmlite==0.40.1
locket==1.0.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==5.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==2.6.3
ninja==1.11.1.1
notebook==6.4.10
numba==0.57.1+1.g1ff679645
numpy==1.24.4
nvfuser==0.1.1+gitunknown
nvidia-dali-cuda120==1.33.0
nvidia-pyindex==1.0.9
nvtx==0.2.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0rc2
opencv==4.7.0
optree==0.10.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pillow==10.2.0
pip==23.3.2
platformdirs==4.1.0
pluggy==1.3.0
ply==3.11
polygraphy==0.49.1
pooch==1.8.0
POT==0.9.4
preshed==3.0.9
prettytable==3.9.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil==5.9.4
ptxcompiler==0.8.1+2.g0d406d6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1.dev0+gba5374836.d20240125
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pybind11-global==2.11.1
pycocotools==2.0+nv0.8.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pykeops==2.2.3
pylibcugraph==23.12.0
pylibcugraphops==23.12.0
pylibraft==23.12.0
pynvml==11.4.1
pyparsing==3.1.1
pytest==7.4.4
pytest-flakefinder==1.1.0
pytest-rerunfailures==13.0
pytest-shard==0.1.2
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-hostlist==1.23.0
pytorch-lightning==2.4.0
pytorch-quantization==2.1.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
raft-dask==23.12.0
rapids-dask-dependency==23.12.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rmm==23.12.0
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.12.0
Send2Trash==1.8.2
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx-glpi-theme==0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
tblib==3.0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorrt==8.6.1
terminado==0.18.0
thinc==8.2.2
threadpoolctl==3.2.0
thriftpy2==0.4.17
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.0a0+81ea7a4
torch-fidelity==0.3.0
torch-tensorrt==2.2.0a0
torchdata==0.7.0a0
torchmetrics==1.4.1
torchtext==0.17.0a0
torchvision==0.17.0a0
tornado==6.4
tqdm==4.66.1
traitlets==5.9.0
transformer-engine==1.2.1+bbafb02
treelite==3.9.1
treelite-runtime==3.9.1
triton==2.1.0+6e4932c
typer==0.9.0
types-dataclasses==0.6.6
typing_extensions==4.9.0
ucx-py==0.35.0
uff==0.6.9
urllib3==1.26.18
wandb==0.17.6
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
xdoctest==1.0.2
xgboost==1.7.6
yarl==1.9.4
zict==3.0.0
zipp==3.17.0
